# internal/spec/swagger.yaml
basePath: /api/v1
definitions:
  github_com_new-marty_health-connect_internal_apperror.Body:
    properties:
      code:
        $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.ErrorCode'
      details:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  github_com_new-marty_health-connect_internal_apperror.ErrorCode:
    enum:
    - NOT_FOUND
    - INVALID_INPUT
    - INTERNAL_ERROR
    - TIMEOUT
    type: string
    x-enum-varnames:
    - CodeNotFound
    - CodeInvalidInput
    - CodeInternal
    - CodeTimeout
  github_com_new-marty_health-connect_internal_apperror.Response:
    properties:
      error:
        $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Body'
    type: object
  internal_analysis.AnalysisRequest:
    properties:
      date:
        type: string
    type: object
  internal_analysis.AnalysisResponse:
    properties:
      analysis:
        type: string
      date:
        type: string
    type: object
...
  internal_hevy.Workout:
    properties:
      bodyweight_kg:
        type: number
      date:
        type: string
      id:
        type: integer
      notes:
        type: string
      session_num:
        type: integer
      session_type:
        type: string
      sets:
        items:
          $ref: '#/definitions/internal_hevy.WorkoutSet'
        type: array
      source:
        type: string
    type: object
  internal_hevy.WorkoutSet:
    properties:
      exercise_name:
        type: string
      id:
        type: integer
      notes:
        type: string
      reps:
        type: integer
      rpe:
        type: number
      set_num:
        type: integer
      weight_kg:
        type: number
      workout_id:
        type: integer
    type: object
  internal_inbody.BodyCompScan:
    properties:
      bfm_kg:
        type: number
      bmi:
        type: number
      bmr_kcal:
        type: integer
      date:
        type: string
      ecw_kg:
        type: number
      ffm_kg:
        type: number
      icw_kg:
        type: number
      mineral_kg:
        type: number
      pbf_pct:
        type: number
      protein_kg:
        type: number
      scan_datetime:
        type: string
      smm_kg:
        type: number
      vfl:
        type: integer
      weight_kg:
        type: number
    type: object
  internal_meals.CreateMealRequest:
    properties:
      calories:
        example: 650
        type: integer
      carbs_g:
        example: 70
        type: number
      date:
        example: "2026-05-04"
        type: string
      description:
        example: chicken rice bowl with broccoli
        type: string
      fat_g:
        example: 18
        type: number
      meal:
        example: lunch
        type: string
      protein_g:
        example: 45
        type: number
      source:
        example: manual
        type: string
    required:
    - date
    - description
    type: object
  internal_meals.Meal:
    properties:
      calories:
        type: integer
      carbs_g:
        type: number
      date:
        type: string
      description:
        type: string
      fat_g:
        type: number
      id:
        type: integer
      meal:
        type: string
      protein_g:
        type: number
      source:
        type: string
    type: object
...
info:
  contact: {}
  description: Personal health data API. Aggregates Oura, Strava, Hevy, InBody, Apple
    Health, and meal logs. Every endpoint is also exposed as an LLM tool — see /openapi.json,
    /tools/openai.json, /tools/anthropic.json.
  title: Health Connect API
  version: "1.0"
paths:
  /analysis:
    post:
      consumes:
      - application/json
      description: Runs a Claude analysis over the daily summary and returns 3–4 specific
        bullet-point insights. Defaults to yesterday if `date` is not provided. Use
        this when the user wants commentary or interpretation rather than raw data.
      operationId: analysis_run
      parameters:
      - description: Analysis request — date is optional and defaults to yesterday
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_analysis.AnalysisRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/internal_analysis.AnalysisResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
      security:
      - BearerAuth: []
      summary: Generate AI health insights for a day
      tags:
      - analysis
...
  /hevy/workouts:
    get:
      description: 'Returns Hevy gym workouts (each with its sets) over an optional
        date range. Use this for strength training questions: volume, sets, reps,
        exercises lifted.'
      operationId: hevy_list_workouts
      parameters:
      - description: Start date (YYYY-MM-DD), inclusive
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD), inclusive
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/internal_hevy.Workout'
                type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
      security:
      - BearerAuth: []
      summary: List Hevy gym workouts
      tags:
      - hevy
  /hevy/workouts/{id}:
    get:
      description: Returns a single gym workout with its full set list.
      operationId: hevy_get_workout
      parameters:
      - description: Hevy workout numeric id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/internal_hevy.Workout'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
      security:
      - BearerAuth: []
      summary: Get one Hevy workout by id
      tags:
      - hevy
  /inbody/scans:
    get:
      description: Returns InBody scans (weight, body fat %, muscle mass, etc.) over
        an optional date range. Use for body composition trends.
      operationId: inbody_list_scans
      parameters:
      - description: Start date (YYYY-MM-DD), inclusive
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD), inclusive
        in: query
        name: to
        type: string
      - description: Maximum number of scans to return (newest first)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/internal_inbody.BodyCompScan'
                type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
      security:
      - BearerAuth: []
      summary: List InBody body composition scans
      tags:
      - inbody
  /inbody/scans/latest:
    get:
      description: Returns the single most recent InBody body composition scan, or
        404 if none exist.
      operationId: inbody_get_latest_scan
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/internal_inbody.BodyCompScan'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
      security:
      - BearerAuth: []
      summary: Get the most recent InBody scan
      tags:
      - inbody
  /meals:
    get:
      description: 'Returns logged meals filtered by an exact date or a date range.
        Use for nutrition questions: calories, macros, what was eaten on a date.'
      operationId: meals_list
      parameters:
      - description: Exact day (YYYY-MM-DD)
        in: query
        name: date
        type: string
      - description: Start date (YYYY-MM-DD), inclusive — ignored if date is set
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD), inclusive — ignored if date is set
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/internal_meals.Meal'
                type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
      security:
      - BearerAuth: []
      summary: List logged meals
      tags:
      - meals
    post:
      consumes:
      - application/json
      description: Creates a meal log entry. Use this when the user wants to record
        what they ate.
      operationId: meals_create
      parameters:
      - description: Meal log entry
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_meals.CreateMealRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
      security:
      - BearerAuth: []
      summary: Log a meal
      tags:
      - meals
  /meals/{id}:
    delete:
      description: Removes a meal entry by id.
      operationId: meals_delete
      parameters:
      - description: Meal id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
      security:
      - BearerAuth: []
      summary: Delete a meal log entry
      tags:
      - meals
...
securityDefinitions:
  BearerAuth:
    description: Bearer token. Set the same value in HEALTH_CONNECT_API_TOKEN on the
      server.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"